New soft-stop function
Added new "soft-stop" function.

Use --stopfile parameter to make it check for the stopfile.txt content before checking if it should buy. This way you can shut down your bot without watching over it when it is still trying to sell. It will finish the last job and then quit.